fastapi/requirements.txt
openai
uvicorn
fastapi
python-dotenv
requests
langchain_community
langchain_core